--- src/types/friend.ts
@@ -1,6 +1,0 @@
-
-export interface Friend {
-    player_id: string
-    name: string
-    smiley_id: number
-}
