--- src/types/world-meta.ts
@@ -1,8 +1,0 @@
-
-export interface WorldMeta {
-    id: string
-    title: string
-    owner: string
-    playCount: number
-    onlineCount: number
-}
